pyproject.toml
[tool.poetry]
name = "ms_analytics"
version = "0.6.2"
description = ""
authors = ["Santicm23 <Santicastrom@outlook.com>"]
readme = "README.md"

[tool.poetry.scripts]
start = "ms_analytics.main:main"
dev = "ms_analytics.main:main(True)"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
